Move splash effect into InWorldRenderer and more javadoc
