+++ src/main/java/org/terasology/gooeyDefence/towerBlocks/targeters/MissileTargeterSystem.java
@@ -20,10 +20,8 @@ import org.terasology.entitySystem.event.ReceiveEvent;
 import org.terasology.entitySystem.systems.RegisterSystem;
 import org.terasology.gooeyDefence.EnemyManager;
 import org.terasology.gooeyDefence.InWorldRenderer;
-import org.terasology.gooeyDefence.components.TargeterBulletComponent;
+import org.terasology.gooeyDefence.components.SplashBulletComponent;
 import org.terasology.gooeyDefence.events.combat.SelectEnemiesEvent;
-import org.terasology.gooeyDefence.movement.components.MovementComponent;
-import org.terasology.gooeyDefence.movement.events.ReachedGoalEvent;
 import org.terasology.logic.location.LocationComponent;
 import org.terasology.math.geom.Vector3f;
 import org.terasology.registry.In;
@@ -66,19 +64,8 @@ public class MissileTargeterSystem extends SniperTargeterSystem {
             inWorldRenderer.shootBulletTowards(
                     target,
                     locationComponent.getWorldPosition(),
-                    new MissileBulletComponent());
+                    new SplashBulletComponent());
         }
         targeterComponent.setLastTarget(target);
-    }
-
-
-    /**
-     * Filters on {@link TargeterBulletComponent}, {@link MissileBulletComponent} and {@link MovementComponent}
-     *
-     * @see ReachedGoalEvent
-     */
-    @ReceiveEvent(components = {TargeterBulletComponent.class, MissileBulletComponent.class})
-    public void onReachedGoal(ReachedGoalEvent event, EntityRef entity, MovementComponent movementComponent) {
-        inWorldRenderer.displayExpandingSphere(movementComponent.getGoal(), 0.5f);
     }
 }
